Split logic to lib.rs
